Move UTF8 string marshalling helpers to shared class
It turns out that Mono sometimes still eagerly calls BeforeFieldInit
cctors. As a result, when the DynDll Windows backend used the stirng
marshalling helpers defined in the Unix interop class, the Unix cctor
would be called, and try to find a working libdl (which obviously
doesn't exist on Windows), and thus fail, throw, and prevent thhe type
from loading properly.
